refactoring of transform
