+++ packages/typia/src/transform.ts
@@ -1,6 +1,5 @@
 import transform from "@typia/transform";
 
-export * from "@typia/transform";
-export * from "./programmers/TypiaProgrammer";
-
 export default transform;
+
+export * from "@typia/transform";
